Split effects.psfs into subpackage
